--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/client/diffsummary/costfunction/CostFunction.java
@@ -1,6 +1,0 @@
-package com.linkedin.thirdeye.client.diffsummary.costfunction;
-
-public interface CostFunction {
-  double getCost(double baselineValue, double currentValue, double parentRatio, double globalBaselineValue,
-      double globalCurrentValue);
-}
