+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/client/diffsummary/costfunctions/CostFunction.java
@@ -1,0 +1,6 @@
+package com.linkedin.thirdeye.client.diffsummary.costfunctions;
+
+public interface CostFunction {
+  double computeCost(double baselineValue, double currentValue, double parentRatio, double globalBaselineValue,
+      double globalCurrentValue);
+}
