+++ thirdeye/thirdeye-pinot/src/test/java/com/linkedin/thirdeye/client/diffsummary/MultiDimensionalSummaryCLIToolTest.java
@@ -1,6 +1,6 @@
 package com.linkedin.thirdeye.client.diffsummary;
 
-import com.linkedin.thirdeye.client.diffsummary.costfunction.BalancedCostFunction;
+import com.linkedin.thirdeye.client.diffsummary.costfunctions.BalancedCostFunction;
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Arrays;
@@ -9,14 +9,14 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 
-public class MultiDimensionalSummaryTest {
+public class MultiDimensionalSummaryCLIToolTest {
   @Test
   public void testInitiateCostFunction()
       throws NoSuchMethodException, IOException, InstantiationException, IllegalAccessException,
              InvocationTargetException, ClassNotFoundException {
     String paramString = String.format("{\"className\":\"%s\"}", BalancedCostFunction.class.getName());
 
-    MultiDimensionalSummary.initiateCostFunction(paramString);
+    MultiDimensionalSummaryCLITool.initiateCostFunction(paramString);
   }
 
   @Test(expectedExceptions = IllegalArgumentException.class)
@@ -25,16 +25,16 @@ public class MultiDimensionalSummaryTest {
              InvocationTargetException, ClassNotFoundException {
     String paramString = "{}";
 
-    MultiDimensionalSummary.initiateCostFunction(paramString);
+    MultiDimensionalSummaryCLITool.initiateCostFunction(paramString);
   }
 
   @Test
   public void testSanitizeDimensions() {
-    List<String> dims = Arrays.asList("environment", "page", "page" + MultiDimensionalSummary.TOP_K_POSTFIX);
+    List<String> dims = Arrays.asList("environment", "page", "page" + MultiDimensionalSummaryCLITool.TOP_K_POSTFIX);
     Dimensions dimensions = new Dimensions(dims);
-    Dimensions sanitizedDimensions = MultiDimensionalSummary.sanitizeDimensions(dimensions);
+    Dimensions sanitizedDimensions = MultiDimensionalSummaryCLITool.sanitizeDimensions(dimensions);
 
-    List<String> expectedSanitizedDims = Arrays.asList("page" + MultiDimensionalSummary.TOP_K_POSTFIX);
+    List<String> expectedSanitizedDims = Arrays.asList("page" + MultiDimensionalSummaryCLITool.TOP_K_POSTFIX);
     Dimensions expectedSanitizedDimensions = new Dimensions(expectedSanitizedDims);
 
     Assert.assertEquals(sanitizedDimensions, expectedSanitizedDimensions);
